requirements.txt
Django==1.5
djangorestframework==2.2.5
MySQL-python==1.2.4
django-admin-bootstrapped==1.6.4
dj-static==0.0.5
Unipath==1.0
dj-database-url==0.2.2
gunicorn==18.0
